+++ src/tests.zig
@@ -130,6 +130,7 @@ test "compile all source files" {
     _ = @import("test/NetworkProbe.zig");
     _ = @import("test/FaultInjection.zig");
     _ = @import("http/StaticFiles.zig");
+    _ = @import("http/Multipart.zig");
     _ = @import("test/ContractGate.zig");
 
     // Tracing